moving code to class
