+++ source/MySEProject/AnomalyDetectionTeamSynergy/RunMultiSequenceLearningExperiment.cs
@@ -1,0 +1,56 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace AnomalyDetectionTeamSynergy
+{
+    internal class RunMultiSequenceLearningExperiment
+    {
+        /// <summary>
+        /// This example demonstrates how to learn two sequences and how to use the prediction mechanism.
+        /// First, two sequences are learned.
+        /// Second, three short sequences with three elements each are created und used for prediction. The predictor used by experiment privides to the HTM every element of every predicting sequence.
+        /// The predictor tries to predict the next element.
+        /// </summary>
+        public static void Run()
+        {
+            Dictionary<string, List<double>> sequences = new Dictionary<string, List<double>>();
+
+            sequences.Add("S1", new List<double>(new double[] { 1, 3, 5, 7, 9, 10 }));
+            sequences.Add("S2", new List<double>(new double[] { 1, 3, 6, 7, 8, 10 }));
+            sequences.Add("S3", new List<double>(new double[] { 3, 4, 5, 7, 9, 11 }));
+
+            MultiSequenceLearning experiment = new MultiSequenceLearning();
+            var predictor = experiment.Run(sequences);
+
+            // These list are used to see how the prediction works.
+            // Predictor is traversing the list element by element. 
+            // By providing more elements to the prediction, the predictor delivers more precise result.
+
+            var list1 = new double[] { 1, 3, 5, 7, 9 };
+
+            foreach (var item in list1)
+            {
+                var res = predictor.Predict(item);
+
+                if (res.Count > 0)
+                {
+                    foreach (var pred in res)
+                    {
+                        Debug.WriteLine($"{pred.PredictedInput} - {pred.Similarity}");
+                    }
+
+                    var tokens = res.First().PredictedInput.Split('_');
+                    var tokens2 = res.First().PredictedInput.Split('-');
+                    Debug.WriteLine($"Predicted Sequence: {tokens[0]}, predicted next element {tokens2.Last()}");
+                }
+                else
+                    Debug.WriteLine("Nothing predicted :(");
+            }
+
+        }
+    }
+}
